+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [2.24.1803.0 - Unreleased]
 ### Added
+- Added Get-PnPHubSite cmdlet
+- Added Set-PnPHubSite cmdlet
+- Added Grant-PnPHubSiteRights cmdlet
 - Added Register-PnPHubSite cmdlet
 - Added Unregister-PnPHubSite cmdlet
 - Added Connect-PnPHubSite cmdlet
@@ -19,6 +22,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Added Get-PnPTenantCdnOrigin cmdlet
 - Added Add-PnPTenantCdnOrigin cmdlet
 - Added Remove-PnPTenantCdnOrigin cmdlet
+- Added Get-PnPTenantCdnPolicies cmdlet
+- Added Set-PnPTenantCdnPolicy cmdlet
 - Added Add-PnPSiteCollectionAppCatalog cmdlet
 - Added Remove-PnPSiteCollectionAppCatalog cmdlet
 - Added Get-PnPNavigationNode cmdlet